Problem: permutations difficult to visualize
The file naming scheme is difficult to decipher, but
even if it is made more descriptive there are many
aspects of the xml and generated code that could not
be captured in the filename.

Solution: use gsl to generate a dot graph of various
relationships and flows. Generate an svg from the dot
file and embed in the local readme. This can be made
more descriptive as needed.